добавлен useFilter, корр. по замечаниями
