+++ project/src/components/app/app.tsx
@@ -19,7 +19,7 @@ function App(): JSX.Element {
   const dispatch = useAppDispatch();
   useEffect(() => {
     dispatch(getOffers({ offers: roomInfos }));
-  }, []);
+  }, [dispatch]);
 
   return (
     <Router>
