+++ project/src/hooks/index.ts
@@ -1,6 +1,10 @@
 import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
+import { useFilter } from './useFilter';
 import type { State, AppDispatch } from '../types/state';
+
 
 export const useAppDispatch = () => useDispatch<AppDispatch>();
 
 export const useAppSelector: TypedUseSelectorHook<State> = useSelector;
+
+export { useFilter };
